requirements.txt
torch
torchvision
numpy
protobuf
PyYAML
scikit-learn
scipy
seaborn
tqdm
wandb
